Прибрался в коде, переместил все обработчики в одно место
