--- cartoonfilter_ds_project/run.bat
@@ -1,1 +1,0 @@
-set FLASK_APP=webapp && set FLASK_ENV=development && set FLASK_DEBUG=1 && flask run
