+++ cartoonfilter_ds_project/run_MacOS_or_Linux.sh
@@ -1,0 +1,2 @@
+#!/bin/sh
+export FLASK_APP=webapp && export FLASK_ENV=development && flask run
