+++ cartoonfilter_ds_project/run_Windows.bat
@@ -1,0 +1,1 @@
+set FLASK_APP=webapp && set FLASK_ENV=development && set FLASK_DEBUG=1 && flask run
