+++ cartoonfilter_ds_project/webapp/config.py
@@ -2,7 +2,7 @@ import os
 
 basedir = os.path.abspath(os.path.dirname(__file__))
 
-SECRET_KEY = ''
+SECRET_KEY = 'Your secret key'
 ALLOWED_EXTENSIONS = {'jpg'}
 SQLALCHEMY_DATABASE_URI = 'sqlite:///' + os.path.join(basedir, '..', 'webapp.db')
